% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_metrics.R
\name{absolute_percent_error}
\alias{absolute_percent_error}
\title{Calculate absolute percent error regression loss.}
\usage{
absolute_percent_error(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}
}
\value{
A numeric vector of length equal to \code{max(length(y_true), length(y_pred))}.
}
\description{
Calculate absolute percent error regression loss.
}

