% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{accuracy_score}
\alias{accuracy_score}
\title{Calculate accuracy score.}
\usage{
accuracy_score(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}
}
\value{
TO BE EDITED.
}
\description{
Calculate accuracy score.
}
\seealso{
Other classification metrics: \code{\link{auc_score}}
}

