% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{ad}
\alias{ad}
\title{Calculate absolute error regression loss.}
\usage{
ad(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}
}
\value{
A numeric vector of length equal to \code{max(length(y_true), length(y_pred))}.
}
\description{
Calculate absolute error regression loss.
}
\seealso{
Other clustering metrics: \code{\link{adjuste}},
  \code{\link{adjust}}, \code{\link{adjus}},
  \code{\link{adju}}
}

