% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_sbi.R
\name{add_distances}
\alias{add_distances}
\title{Adding the distance-based metric}
\usage{
add_distances(.data, pareto_solutions = NULL, parallel_cores = 0)
}
\arguments{
\item{.data}{The optimization data to use, should be imported with \code{\link[=load_dataset]{load_dataset()}}}

\item{pareto_solutions}{Iteration number to use as interpolant, default NULL}

\item{parallel_cores}{Number of cores to use for calculation. Default 0, a positive value will enable parallelization.}
}
\description{
Apply the distance-based metric, with a new column Distance
Find the minimum RSME to a Pareto-optimal solution
Pareto-optimal solutions have Rank 1, placed in Y
}
