% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{ae}
\alias{ae}
\title{Calculate absolute error between actual and forecast.}
\usage{
ae(y_true = NULL, y_pred = NULL, actual = NULL, forecast = NULL)
}
\arguments{
\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}

\item{actual}{a numeric vector of actuals.}

\item{forecast}{a numeric vector of forecasts.}
}
\value{
a numeric vector of absolute errors.
}
\description{
\code{ae} takes actual and forecast numeric vectors and returns a
numeric vector where forecast is subtracted from the actual and
then the absolute value is taken of those errors.
}
\examples{
# Examples
ae(1:100, 100:1)
}

