% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ape}
\alias{ape}
\title{Calculate absolute percent error between actual and forecast.}
\usage{
ape(actual, forecast)
}
\arguments{
\item{actual}{A numeric vector of actuals.}

\item{forecast}{A numeric vector of forecasts.}
}
\value{
A numeric vector of absolute percent errors.
}
\description{
\code{pe} takes actual and forecast numeric vectors and returns a
numeric vector where forecast is subtracted from the actual and
then those errors are divided by the actuals. Lastly, the
absolute value of those percent errors are taken.
}
\examples{
# Examples
ape(1:100, 100:1)
}

