% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpm.R
\name{assign_reference_point}
\alias{assign_reference_point}
\title{assign reference point}
\usage{
assign_reference_point(.data, reference_point, knn)
}
\arguments{
\item{.data}{data to analyze}

\item{reference_point}{named list of reference points}

\item{knn}{the fraction of neighbouring solutions to include}
}
\description{
assign reference point
}
