% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/errors.R
\name{auc}
\alias{auc}
\title{Calculate area under the curve between actual and forecast.}
\usage{
auc(actual, forecast)
}
\arguments{
\item{actual}{A binary numeric vector of actuals.}

\item{forecast}{A binary numeric vector of forecasts.}
}
\value{
A numeric vector of length one.
}
\description{
\code{auc} calculates the area under the curve.
}
\examples{
# Examples
auc(1:100, 100:1)
}

