% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{auc_score}
\alias{auc_score}
\title{Calculate area under the curve.}
\usage{
auc_score(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}
}
\value{
TO BE EDITED.
}
\description{
Calculate area under the curve.
}
\seealso{
Other classification metrics: \code{\link{accuracy_score}}
}

