% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ce}
\alias{ce}
\title{Calculate classification error between actual and forecast.}
\usage{
ce(actual, forecast)
}
\arguments{
\item{actual}{A numeric vector of actuals.}

\item{forecast}{A numeric vector of forecasts.}
}
\value{
A numeric vector of errors.
}
\description{
\code{e} takes actual and forecast numeric vectors and returns the
classification error.
}
\examples{
# Examples
ce(1:100, 100:1)
}

