% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{e}
\alias{e}
\title{Calculate error between actual and forecast.}
\usage{
e(y_true = NULL, y_pred = NULL, actual = NULL, forecast = NULL)
}
\arguments{
\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}

\item{actual}{a numeric vector of actuals.}

\item{forecast}{a numeric vector of forecasts.}
}
\value{
a numeric vector of errors.
}
\description{
\code{e} takes actual and forecast numeric vectors and returns a
numeric vector where forecast is subtracted from the actual.
}
\examples{
# Examples
e(1:100, 100:1)
}

