% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_metrics.R
\name{explained_variance_score}
\alias{explained_variance_score}
\title{Calculate explained variance regression score function.}
\usage{
explained_variance_score(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}
}
\value{
A numeric vector of length one.
}
\description{
Calculate explained variance regression score function.
}
\examples{
explained_variance_score(1:10, 10:1)
}
\seealso{
Other regression.metrics: \code{\link{absolute_error}},
  \code{\link{absolute_percent_error}},
  \code{\link{log_error}},
  \code{\link{mean_absolute_error}},
  \code{\link{mean_absolute_percent_error}},
  \code{\link{mean_absolute_scaled_error}},
  \code{\link{mean_error}},
  \code{\link{mean_percent_error}},
  \code{\link{mean_squared_error}},
  \code{\link{median_absolute_error}},
  \code{\link{median_absolute_percent_error}},
  \code{\link{median_percent_error}},
  \code{\link{median_squared_error}},
  \code{\link{percent_error}}, \code{\link{r2_score}},
  \code{\link{squared_error}},
  \code{\link{total_variance_score}},
  \code{\link{unexplained_variance_score}}
}

