% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SCORE.R
\name{export_plot}
\alias{export_plot}
\title{Exports a plot created by \code{\link[=freqchart]{freqchart()}} or \code{\link[=plot_pareto]{plot_pareto()}}}
\usage{
export_plot(
  .plot = ggplot2::last_plot(),
  filename,
  path = getwd(),
  device = "pdf",
  width = 15,
  height = 7,
  ...
)
}
\arguments{
\item{filename}{The name of the resulting file}

\item{device}{png, pdf, or svg to ggsave or tex for tikz. Supply vector to save as multiple files.}
}
\description{
Exports a plot created by \code{\link[=freqchart]{freqchart()}} or \code{\link[=plot_pareto]{plot_pareto()}}
}
