% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpm.R
\name{fpm}
\alias{fpm}
\title{Flexible Pattern Mining FPM}
\usage{
fpm(
  .data,
  maxLevel = 1,
  minSig = 0.5,
  selectedData,
  unselectedData = NULL,
  useEquality = TRUE,
  onlyMostSignificant = TRUE
)
}
\arguments{
\item{.data}{Data input, obtained using \code{\link[=loaddataset]{loaddataset()}}}
}
\value{
A list of lists where each combination of itemsets are considered
}
\description{
Flexible pattern mining in R(), returns rules to reach certain areas in the objectives space
}
