% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{freqchart}
\alias{freqchart}
\title{Creates a frequency chart from the supplied data set}
\usage{
freqchart(
  .data,
  objectives = names(attr(.data, "objectives")),
  rank = 1,
  n = 10
)
}
\arguments{
\item{data}{A dataset loaded with \code{\link[=loaddataset]{loaddataset()}}, and has ranks applied through \code{\link[=ndsecr]{ndsecr()}}}
}
\value{
A ggplot2 object showing the frequencies of parameters
}
\description{
Creates a frequency chart from the supplied data set
}
