% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_dataset}
\alias{load_dataset}
\title{Load optimization data}
\usage{
load_dataset(
  filename,
  objectives = NULL,
  inputs = NULL,
  outputs = NULL,
  custom_outputs = NULL
)
}
\arguments{
\item{filename}{A csv file containing formatted optimization data from Opt}

\item{objectives}{A character vector naming the objectives}

\item{inputs}{A character vector naming the inputs}

\item{outputs}{A character vector naming the outputs}

\item{custom_outputs}{A character vector describing custom outputs}
}
\value{
A dataframe with the data
}
\description{
Load optimization data
}
