% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loaddataset}
\alias{loaddataset}
\title{Loads optimization data and returns an OptResult class.}
\usage{
loaddataset(filename, objectives, outputs, custom_outputs)
}
\arguments{
\item{file}{A file containing optimization data}
}
\value{
A dataframe with the data
}
\description{
Loads optimization data and returns an OptResult class.
}
