% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/errors.R
\name{lse}
\alias{lse}
\title{Calculate log squared errors between actual and forecast.}
\usage{
lse(actual, forecast)
}
\arguments{
\item{actual}{A numeric vector of actuals.}

\item{forecast}{A numeric vector of forecasts.}
}
\value{
A numeric vector.
}
\description{
Calculate log squared errors between actual and forecast.
}
\examples{
# Examples
lse(1:100, 100:1)
}

