% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{matthews_corrcoef}
\alias{matthews_corrcoef}
\title{TO BE EDITED}
\usage{
matthews_corrcoef(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}
}
\value{
TO BE EDITED.
}
\description{
TO BE EDITED
}
\seealso{
Other classification metrics: \code{\link{accuracy_score}},
  \code{\link{auc_score}},
  \code{\link{average_precision_score}},
  \code{\link{brier_score}},
  \code{\link{classification_report}},
  \code{\link{confusion_matrix}}, \code{\link{f1_score}},
  \code{\link{fbeta_score}}, \code{\link{hamming_loss}},
  \code{\link{hinge_loss}},
  \code{\link{jaccard_similarity_score}},
  \code{\link{log_loss}},
  \code{\link{precision_recall_curve}},
  \code{\link{precision_recall_fscore_support}},
  \code{\link{precision_score}},
  \code{\link{recall_score}}, \code{\link{roc_auc_score}},
  \code{\link{roc_curve}}, \code{\link{zero_one_loss}}
}

