% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mean_absolute_error}
\alias{mean_absolute_error}
\title{Calculate mean absolute error regression loss.}
\usage{
mean_absolute_error(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}
}
\value{
A numeric vector of length equal to \code{max(length(y_true), length(y_pred))}.
}
\description{
Calculate mean absolute error regression loss.
}

