% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_metrics.R
\name{mean_absolute_percent_error}
\alias{mean_absolute_percent_error}
\title{Calculate mean absolute percent error regression loss.}
\usage{
mean_absolute_percent_error(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}
}
\value{
A numeric vector of length one.
}
\description{
Calculate mean absolute percent error regression loss.
}
\examples{
mean_percent_error(1:10, 10:1)
}

