% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_metrics.R
\name{mean_error}
\alias{mean_error}
\title{Calculate mean error regression loss.}
\usage{
mean_error(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}
}
\value{
A numeric vector of length one.
}
\description{
Calculate mean error regression loss.
}

