% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{median_absolute_error}
\alias{median_absolute_error}
\title{Calculate median absolute error regression loss.}
\usage{
median_absolute_error(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}
}
\value{
A numeric vector of length one.
}
\description{
Calculate median absolute error regression loss.
}

