% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{median_percent_error}
\alias{median_percent_error}
\title{Calculate median percent error regression loss.}
\usage{
median_percent_error(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}
}
\value{
A numeric vector of length one.
}
\description{
Calculate median percent error regression loss.
}

