% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_metrics.R
\name{median_squared_error}
\alias{median_squared_error}
\title{Calculate median squared error regression loss.}
\usage{
median_squared_error(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}
}
\value{
A numeric vector of length one.
}
\description{
Calculate median squared error regression loss.
}

