% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{median_squared_log_error}
\alias{median_squared_log_error}
\title{Calculate median squared log error regression loss.}
\usage{
median_squared_log_error(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}
}
\value{
A numeric vector of length one.
}
\description{
Calculate median squared log error regression loss.
}
\examples{
median_squared_log_error(1:10, 10:1)
}
\seealso{
Other regression metrics: \code{\link{absolute_error}},
  \code{\link{absolute_percent_error}},
  \code{\link{explained_variance_score}},
  \code{\link{log_error}},
  \code{\link{mean_absolute_error}},
  \code{\link{mean_absolute_percent_error}},
  \code{\link{mean_absolute_scaled_error}},
  \code{\link{mean_error}},
  \code{\link{mean_percent_error}},
  \code{\link{mean_squared_error}},
  \code{\link{mean_squared_log_error}},
  \code{\link{median_absolute_error}},
  \code{\link{median_absolute_percent_error}},
  \code{\link{median_percent_error}},
  \code{\link{median_squared_error}},
  \code{\link{percent_error}}, \code{\link{r2_score}},
  \code{\link{squared_error}},
  \code{\link{squared_log_error}},
  \code{\link{symmetric_mean_absolute_percent_error}},
  \code{\link{symmetric_median_absolute_percent_error}},
  \code{\link{total_variance_score}},
  \code{\link{unexplained_variance_score}}
}

