% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/errors.R
\name{mlse}
\alias{mlse}
\title{Calculate mean log squared errors between actual and forecast.}
\usage{
mlse(actual, forecast)
}
\arguments{
\item{actual}{A numeric vector of actuals.}

\item{forecast}{A numeric vector of forecasts.}
}
\value{
A numeric vector of length one.
}
\description{
Calculate mean log squared errors between actual and forecast.
}
\examples{
# Examples
msle(1:100, 100:1)
}

