% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/errors.R
\name{mse}
\alias{mse}
\title{Calculate mean squared errors between actual and forecast.}
\usage{
mse(actual, forecast)
}
\arguments{
\item{actual}{A numeric vector of actuals.}

\item{forecast}{A numeric vector of forecasts.}
}
\value{
A numeric vector of length one.
}
\description{
\code{mse} calculates the mean squared errors between actual and forecast.
}
\examples{
# Examples
mse(1:100, 100:1)
}

