% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/errors.R
\name{msle}
\alias{msle}
\title{Title.}
\usage{
msle(actual, forecast)
}
\arguments{
\item{actual}{A numeric vector of actuals.}

\item{forecast}{A numeric vector of forecasts.}
}
\value{
A numeric vector of length one.
}
\description{
\code{function_name} does something.
}
\examples{
# Examples
}

