% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nds.R
\name{ndsecr}
\alias{ndsecr}
\title{Perform non-dominated sorting}
\usage{
ndsecr(.data, objectives = .data$objectives)
}
\arguments{
\item{objectives}{objectives from optresults or an character/numeric vector
defining the objectives to subset}

\item{optresults}{an OptResults class}
}
\value{
optresults with ranks
}
\description{
Perform non-dominated sorting
}
