% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nds.R
\name{normalize_values}
\alias{normalize_values}
\title{Normalize values in the optimization data.}
\usage{
normalize_values(.data, objectives = attr(.data, "objectives"), limits = NULL)
}
\arguments{
\item{.data}{an OptResults class}

\item{objectives}{Optimization goals}

\item{limits}{Named character vector with min and max values for each objective}
}
\value{
Normalized values
}
\description{
Normalize values in the optimization data.
}
