% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nds.R
\name{normalizeValues}
\alias{normalizeValues}
\title{Normalize values in the optimization data.}
\usage{
normalizeValues(optData, optGoals)
}
\arguments{
\item{optData}{an OptResults class}

\item{optGoals}{Optimization goals}
}
\value{
Normalized values
}
\description{
Normalize values in the optimization data.
}
