% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOREr.R
\name{optresult}
\alias{optresult}
\title{An S3 class representing the optimization result, usually returned from \code{\link[=load_dataset]{load_dataset()}}}
\usage{
optresult(data, opt_name, opt_id = 1, objectives, inputs, outputs, parameters)
}
\arguments{
\item{data}{A tibble containing optimization results from OptimizeBrowser}

\item{opt_name}{A name for the optimization run}

\item{objectives}{A named character vector}

\item{inputs}{A character vector with names of the inputs}

\item{outputs}{A character vector with names of the outputs}

\item{id}{id number of the optimization}
}
\value{
An S3 class
}
\description{
An S3 class representing the optimization result, usually returned from \code{\link[=load_dataset]{load_dataset()}}
}
