% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sbi.R
\name{plotnd}
\alias{plotnd}
\title{Plot multi-objective optimization data}
\usage{
plotnd(.data, ...)
}
\arguments{
\item{.data}{The dataset generated by \code{\link[=load_dataset]{load_dataset()}}.}

\item{...}{Additional parameters to the function. Possible to define x, y, z, and color parameters.
If the dimensions are not set, they are read from .data.
Non-matched arguments will be further piped to \code{\link[plotly:plot_ly]{plotly::plot_ly()}}.}
}
\description{
Plot data in either two or three dimensions using plotly.
}
\examples{
# Plot in three dimensions with "Rank" as color parameter and
# default axis read from .data$objective_names.
plotnd(SCORER::FMC)

# All objectives defined, along with color parameter.
plotnd(SCORER::FMC, x = "TP", y = "Investment", z = "LeanBuffer", color = "Iteration")

}
\author{
Simon Lidberg
}
