% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scorer.R
\docType{package}
\name{scorer}
\alias{scorer}
\alias{scorer-package}
\title{scorer: Quickly Score Models.}
\description{
A set of tools for quickly scoring models in data science and machine
learning. This toolset is written in C++ for blazing fast performance.
}

