% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOREr.R
\name{select.OptResult}
\alias{select.OptResult}
\title{Make sure to update the objective lists when adding or removing columns.}
\usage{
\method{select}{OptResult}(data, ...)
}
\description{
Make sure to update the objective lists when adding or removing columns.
}
