% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_metrics.R
\name{total_variance_score}
\alias{total_variance_score}
\title{Calculate total variance regression score function.}
\usage{
total_variance_score(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}
}
\value{
A numeric vector of length one.
}
\description{
Calculate total variance regression score function.
}
\examples{
total_variance_score(1:10, 10:1)
}

