% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{true_positive_rate}
\alias{true_positive_rate}
\title{TO BE EDITED}
\usage{
true_positive_rate(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}
}
\description{
TO BE EDITED
}
\seealso{
Other classification metrics: \code{\link{accuracy_score}},
  \code{\link{accuracy}}, \code{\link{condition_negative}},
  \code{\link{condition_positive}},
  \code{\link{diagnostic_odds_ratio}},
  \code{\link{f1_score}},
  \code{\link{false_discovery_rate}},
  \code{\link{false_negative}},
  \code{\link{false_omission_rate}},
  \code{\link{false_positive}},
  \code{\link{matthews_corrcoef}},
  \code{\link{negative_likelihood_ratio}},
  \code{\link{negative_predictive_value}},
  \code{\link{positive_likelihood_ratio}},
  \code{\link{positive_predictive_value}},
  \code{\link{precision}},
  \code{\link{predicted_condition_negative}},
  \code{\link{predicted_condition_positive}},
  \code{\link{prevalence}}, \code{\link{recall}},
  \code{\link{sensitivity}}, \code{\link{specificity}},
  \code{\link{total_population}},
  \code{\link{true_negative_rate}},
  \code{\link{true_negative}}, \code{\link{true_positive}}
}

