% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ae}
\alias{ae}
\title{Calculate absolute error between actual and forecast.}
\usage{
ae(actual, forecast)
}
\arguments{
\item{actual}{A numeric vector of actuals.}

\item{forecast}{A numeric vector of forecasts.}
}
\value{
A numeric vector of absolute errors.
}
\description{
\code{ae} takes actual and forecast numeric vectors and returns a
numeric vector where forecast is subtracted from the actual and
then the absolute value is taken of those errors.
}
\examples{
# Examples
ae(1:100, 100:1)
}

