% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{check_args}
\alias{check_args}
\title{TO BE EDITED}
\usage{
check_args(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}
}
\description{
TO BE EDITED
}

