% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{e}
\alias{e}
\title{Calculate error between actual and forecast.}
\usage{
e(actual, forecast)
}
\arguments{
\item{actual}{A numeric vector of actuals.}

\item{forecast}{A numeric vector of forecasts.}
}
\value{
A numeric vector of errors.
}
\description{
\code{e} takes actual and forecast numeric vectors and returns a
numeric vector where forecast is subtracted from the actual.
}
\examples{
# Examples
e(1:100, 100:1)
}

