% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_metrics.R
\name{explained_variance_score}
\alias{explained_variance_score}
\title{Calculate explained variance regression score function.}
\usage{
explained_variance_score(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}
}
\value{
A numeric vector of length one.
}
\description{
Calculate explained variance regression score function.
}

