% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SCORE.R
\name{freqchart}
\alias{freqchart}
\title{Creates a frequency chart from the supplied data set}
\usage{
freqchart(
  .data,
  objectives = .data$objective_names,
  inputs = .data$inputs,
  rank = 1,
  n = 10
)
}
\arguments{
\item{.data}{The data, preferably loaded from \code{\link[=load_dataset]{load_dataset()}}}

\item{objectives}{Which objectives to use, defaults to all}

\item{inputs}{The inputs to use, defaults to all optimization inputs}

\item{rank}{The rank of the included solutions, inclusive}

\item{n}{The number of bottlenecks to include}
}
\value{
A ggplot2 object showing the frequencies of parameters
}
\description{
Creates a frequency chart from the supplied data set
}
