% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mae}
\alias{mae}
\title{Calculate mean absolute error between actual and forecast.}
\usage{
mae(actual, forecast)
}
\arguments{
\item{actual}{A numeric vector of actuals.}

\item{forecast}{A numeric vector of forecasts.}
}
\value{
A numeric vector of length one representing the mean of absolute errors.
}
\description{
\code{mae} takes actual and forecast numeric vectors and returns a
numeric vector where forecast is subtracted from the actual, the
absolute value is taken of those errors, and lastly, the mean is
taken of those absolute erros.
}
\examples{
# Examples
mae(1:100, 100:1)
}

