% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOREr.R
\name{optresult}
\alias{optresult}
\title{Title}
\usage{
optresult(data, opt_name, id, objectives, inputs, outputs)
}
\arguments{
\item{data}{A tibble containing optimization results from OptimizeBrowser}

\item{opt_name}{A name for the optimization run}

\item{id}{id number of the optimization}

\item{objectives}{A named character vector}

\item{inputs}{A character vector with names of the inputs}

\item{outputs}{A character vector with names of the outputs}
}
\value{
An S3 class
}
\description{
Title
}
