% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{pe}
\alias{pe}
\title{Calculate percent error between actual and forecast.}
\usage{
pe(actual, forecast)
}
\arguments{
\item{actual}{a numeric vector of actuals.}

\item{forecast}{a numeric vector of forecasts.}
}
\value{
a numeric vector of percent errors.
}
\description{
\code{pe} takes actual and forecast numeric vectors and returns a
numeric vector where forecast is subtracted from the actual and
then those errors are divided by the actuals.
}
\examples{
# Examples
pe(1:100, 100:1)
}

