% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_metrics.R
\name{percent_error}
\alias{percent_error}
\title{Calculate percent error regression loss.}
\usage{
percent_error(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}
}
\value{
A numeric vector of length equal to \code{max(length(y_true), length(y_pred))}.
}
\description{
Calculate percent error regression loss.
}
\examples{
percent_error(1:10, 10:1)
}

