% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_metrics.R
\name{r2_score}
\alias{r2_score}
\title{Calculate R^2 (coefficient of determination) regression score function.}
\usage{
r2_score(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}
}
\value{
A numeric vector of length one.
}
\description{
Calculate R^2 (coefficient of determination) regression score function.
}
\examples{
r2_score(1:10, 10:1)
}

