% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmse}
\alias{rmse}
\title{Calculate root mean squared error.}
\usage{
rmse(actual, forecast)
}
\arguments{
\item{actual}{A numeric vector of actuals.}

\item{forecast}{A numeric vector of forecasts.}
}
\value{
A numeric vector of length one: the root mean square of errors.
# Examples
rmse(1:100, 100:1)
}
\description{
\code{rmse} takes actual and forecast numeric vectors and returns a
numeric vector where forecast is subtracted from the actual,
the errors are then squared, the average of those squared error is
taken, and lastly, the root taken of that mean squared error value.
}

