% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scorer.R
\docType{package}
\name{scorer}
\alias{scorer}
\alias{scorer-package}
\title{scorer: Quickly Score Models.}
\description{
The scorer package provides convenience functions for
quickly scoring models.
}
\details{
To learn more about scorer, start with the vignettes:
\code{browseVignettes(package = "scorer")}
}

